pyproject.toml
[project]
name = "crs-zarrpipelines"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.9.13"
dependencies = [
    "bokeh>=3.4.3",
    "cloudpickle>=3.1.2",
    "coiled>=1.132.0",
    "dask>=2024.8.0",
    "exactextract>=0.2.0",
    "fastapi>=0.115.0",
    "google-cloud-run>=0.10.0",
    "fsspec>=2024.6.0",
    "gcsfs>=2025.10.0",
    "geopandas>=1.0.1",
    "jupyterlab>=4.5.5",
    "matplotlib>=3.9.4",
    "msgpack>=1.1.2",
    "numpy>=2.0.2",
    "pandas>=2.3.3",
    "prefect>=3.4.25",
    "pyarrow>=21.0.0",
    "python-dotenv>=1.0",
    "pyyaml>=6.0.3",
    "rasterstats>=0.20.0",
    "requests>=2.32.5",
    "rioxarray>=0.15.0",
    "spyder>=6.1.3",
    "tornado>=6.5.4",
    "uvicorn>=0.30.0",
    "xarray>=2024.7.0",
    "xvec>=0.3.1",
    "zarr>=2.18.2",
]

[dependency-groups]
dev = [
    "pytest>=8.0",
]

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["app"]

[tool.hatch.build.targets.wheel.force-include]
"app/config/pipeline.yaml" = "app/config/pipeline.yaml"
"app/config/scoring.yaml" = "app/config/scoring.yaml"
